python-backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pdfplumber>=0.9.0
python-docx>=0.8.11
openai>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6